ref(sdk-crashes): Rename crash_event fixture to crash_event_cocoa
The crash_event test fixture only contains Cocoa SDK crashes,
and therefore can be renamed to crash_event_cocoa.
