+++ fixtures/sdk_crash_detection/crash_event_cocoa.py
